+++ DLPrototype/Views/Planning/Notes.swift
@@ -1,0 +1,28 @@
+//
+//  Notes.swift
+//  DLPrototype
+//
+//  Created by Ryan Priebe on 2023-08-30.
+//  Copyright © 2023 YegCollective. All rights reserved.
+//
+
+import SwiftUI
+
+extension Planning {
+    struct Notes: View {
+        public var notes: FetchedResults<Note>
+        public var colour: Color
+
+        @EnvironmentObject public var nav: Navigation
+
+        var body: some View {
+            if notes.count > 0 {
+                VStack(alignment: .leading, spacing: 0) {
+                    ForEach(notes) { note in
+                        Row(note: note, colour: colour)
+                    }
+                }
+            }
+        }
+    }
+}
